Remove resolve member function, see commit body
The resolve member function of the Challenge struct made state changes
which would result in only the partial completion of the resolution
task. This was more dangerous than the code it de-duped, because it
made it seem that other *necessary* state changes that might be required
in the complete resolution of a challenge might be done by the resolve
method. They're not. Challenge resolutions are heterogenous, and short
of being able to pass function types into a resolve method, the resolver
itself needs to operate on the internal state of the Challenge struct.
More important to keep that totally explicit than it is to dedupe a
small number of lines. (Though if, in the future, we could solve this
more elegantly that would be highly desirable).