Address audit finding 3.2.
Previously, PLCRVoting checked for the presence of commit hashes in
specific DLL nodes to determine whether a user had participated in those
polls. If a user provided a commit hash of zero for some poll, however,
various checks in the contract relying on the presence of a non-zero
commit hash at a poll to determine user participation would incorrectly
fail.

Relying on user input to make determinations about contract state is
risky in general, so this commit adds explicit flags not controlled by
the user to make determinations where commit hashes had been used
previously.

* Add didCommit and didReveal mappings to poll structs
* In rescueTokens, check that the user's DLL contains the given pollID
  rather than checking if the user had revealed for that poll. Also
  check that the poll has ended. If the poll has ended and the user's
  DLL contains the poll, the user committed but did not reveal.
* In commitVote, when checking whether the provided _prevPollID is
  valid, use the DLL `contains` method instead of checking for a commit
  hash at the _prevPollID index.
* At the end of commitVote, set didCommit for this message sender in
  this poll to true.
* In revealVote, use the didCommit and didReveal flags to determine
  whether the user can reveal for this poll.
* Add brackets to the if-else clause in reveal vote.
* At the end of reveal vote, set didReveal for this message sender in
  this poll to true.
* In getNumPassingTokens, use the didReveal flag to determine whether
  the user revealed for the provided poll.
* Remove the hasBeenRevealed function.