Address audit finding 3.6
Previously, it was possible to process reparameterizations which would
set the dispensation percentage parameters to values greater than 100.
Because the dispensation percentage parameters are meant to be
percentages, this could cause unexpected behaviors in the contracts
which could result in some tokens being locked.

Parameterizations in the disallowed set are those for dispensationPct
and pDispensationPct whose proposed values are greater than 100.

* Check input values in proposeReparameterization and revert if they
  are in the disallowed set.
* Defensive programming: add assertions to processProposal preventing
  such reparameterizations.
* Defensive programming: use SafeMath in challengeReparameterization
  preventing underflows which could have resulted from such
  reparameterizations.
* Add a test to proposeReparameterization asserting that
  parameterizations in the disallowed set will revert.