Address audit finding 3.8.
The appExpiry and processBy values for reparameterization proposals
can overflow if pApplyStageLen, pCommitStageLen, pRevealStageLen, or
any combination thereof, are parameterized to very large numbers.

* Add test reproducing audit finding 3.8.
* Adds check for overflows of appExpiry and processBy in
  processProposal
* Defensive programming: Use SafeMath in calculations in
  proposeReparameterization

Note: proposals which would cause overflows can never be processed when
a challenge against them fails. The proposal will be orphaned and the
proposer will neither be able to set the parameter nor collect their
challenge reward. Furthermore, the voters who voted in the challenge
will not be able to collect their rewards.