Address audit findings 3.11.
The behavior for resolving challenges when nobody participates in the
vote is inconsistently documented. The implemented (and correct)
behavior in both the Registry and Parameterizer is that nobody voting
is handled as a tie insofar as the challenger wins. Furthermore, the
challenger receives the candidate's entire deposit in this case, since
otherwise tokens would be reserved for non-existant voters.

* Make documentation corrections regarding challenge results when
  nobody voted.