Add a tokenClaims getter to Parameterizer.sol
It is not possible to lookup the status of whether voters have claimed
tokens in a challenge from outside the contract, because solc does not
generate getters for nested data structures such as the tokenClaims
mapping in the parameterizer's challenge struct. A tokenClaims getter
such as that in Registry.sol should be added to Parameterizer.sol such
that such lookups can be done.

* Add a tokenClaims getter to Parameterizer.sol
* Add two tests in tokenClaims.js, asserting that the getter returns
  false for a user and a specific challenge before that user has claimed
  their voting rewards, and true after they have done so.