Make factory tests more explicit
Some of the factory tests asserted that the creator of a contract was
account[0], so one thing this commit does is explicitly send contract
creation messages from that account.

In newRegistryBYOToken.js, it appears as though the test was
overspecified, and tested things not actually related to the function
purported as being under test. This commit removes the extraneous code.